Added more code coverage.
